// common/MISA.QLCH.CommonStores/src/models/employee/IEmployee.ts
export default interface IEmployee {

    CodeEmployee?: string;
    NameEmployee?: string;
    GenderEmployee?: string;
    BirthEmployee?: Date;
    TelephoneEmployee?: number;
    StatusEmployee?: string;
    PasswordEmployee?: string;
    IdCardEmployee?: number;
    DateCardEmployee?: Date;
    AddressCardEmployee?: string;
    MaritalEmployee?: string;
    ConfirmPasswordEmployee?: string;
    UsageSystem?: string;
    HomePhone?: number
    AddressHome?: string;
    NationEmployee?: string;
    DistrictEmployee?: string;
    ProvinceEmployee?: string;
    WardsEmployee?: string;
    AddressPresent?: string;
    NationPresent?: string;
    DistrictPresent?: string;
    ProvincePresent?: string;
    WardsPresent?: string;
    EmergencyEmployee?: string;
    RelationshipEmployee?: string;
    PhoneEmergency?: number;
    LandlinePhone?: number;
    EmailHome?: string;
    AddressEmergency?: string;

}